web-app 2.3 => 4.0, JSTL 및 EL 도입해 자바 문법 제거

+++ src/main/webapp/jsp/article/list.jsp
@@ -1,13 +1,6 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
-<%@ page import="java.util.List"%>
-<%@ page import="java.util.Map"%>
-<%@ page import="com.ukj.exam.dto.Article"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 
-<%
-  List<Article> articles = (List<Article>) request.getAttribute("articles");
-  int cPage = (int) request.getAttribute("page");
-  int totalPage = (int) request.getAttribute("totalPage");
-%>
 <!doctype html>
 <html lang="ko">
 <head>
@@ -16,7 +9,9 @@
 </head>
 <body>
   <h1>게시물 리스트</h1>
+
   <%@ include file="../part/topBar.jspf"%>
+
   <div>
     <a href="../home/main">홈으로 이동</a>
     <a href="../article/write">게시물 작성</a>
@@ -37,18 +32,18 @@
       </tr>
     </thead>
     <tbody>
-    <% for (Article article : articles) { %>
+    <c:forEach items="${articles}" var="article">
     <tr>
-      <td><%=article.id%></td>
-      <td><%=article.regDate%></td>
-      <td><%=article.updateDate%></td>
-      <td> <a href="detail?id=<%=article.id%>"><%=article.title%></a></td>
+      <td>${article.id}</td>
+      <td>${article.regDate}</td>
+      <td>${article.updateDate}</td>
+      <td> <a href="detail?id=${article.id}">${article.title}</a></td>
       <td>
-        <a href="doDelete?id=<%=article.id%>">삭제</a>
-        <a href="modify?id=<%=article.id%>">수정</a>
+        <a onclick="if (confirm('삭제하시겠습니까?') == false) return false;" href="doDelete?id=${article.id}">삭제</a>
+        <a href="modify?id=${article.id}">수정</a>
       </td>
     </tr>
-    <% } %>
+    </c:forEach>
     </tbody>
   </table>
   <style type="text/css">
@@ -57,24 +52,30 @@
     }
   </style>
   <div class="page">
-    <%
-      if (cPage > 1) { %>
-        <a href="list?page=1">◄</a>
-    <% }
+    <%-- 원래는 이렇게 가져왔지만 EL을 써서 편하게 가져옴 --%>
+    <%-- int cPage = (int) request.getAttribute("page"); --%>
+    <c:set var="cPage" value="${page}"/>
+    <c:set var="totalPage" value="${totalPage}"/>
+    <c:set var="pageMenuSize" value="5"/>
+    <c:set var="from" value="${cPage - pageMenuSize}"/>
+    <c:set var="end" value="${cPage + pageMenuSize}"/>
 
-      int pageMenuSize = 5;
-      int from = cPage - pageMenuSize;
-      int end = cPage + pageMenuSize;
+    <c:if test="${cPage > 1}">
+      <a href="list?page=1">◄</a>
+    </c:if>
 
-      if ( from < 1 ) { from = 1; }
-      if ( end > totalPage ) { end = totalPage; }
-      for (int i = from; i <= end; i++) {
-    %>
-      <a class="<%=cPage == i ? "red" : ""%>" href="list?page=<%= i%>"><%= i%></a>
-    <% }
-      if (cPage < totalPage) { %>
-        <a href="../article/list?page=<%=totalPage%>">►</a>
-    <% } %>
+    <c:set var="start" value="${from < 1 ? 1 : from}" />
+    <c:set var="end" value="${end > totalPage ? totalPage : end}"/>
+
+    <c:forEach var="i" begin="${start}" end="${end}" step="1">
+      <c:set var="aClassRed" value="${cPage == i ? 'red' : ''}"/>
+      <a class="${aClassRed}" href="list?page=${i}">${i}</a>
+    </c:forEach>
+
+
+    <c:if test="${cPage < totalPage}">
+      <a href="../article/list?page=${totalPage}">►</a>
+    </c:if>
   </div>
 </body>
 </html>
